util: fix UB in float2fixed with negative numbers.
For negative values inside the valid range, float2fixed simply
multiplies that value by the fixed point factor, resulting in another
negative floating point value. Assigning that value to an unsigned
integer is undefined behavior, since it's a value outside of its range.

In order to fix this, we can simply add an intermediate cast through
int32_t, which can represent negative values, because that's the
underlying type for this value; the function only returns uint32_t to
simplify any further register field manipulation.

Reported-by: UndefinedBehaviorSanitizer

+++ util/util.cc
@@ -45,7 +45,7 @@ uint32_t float2fixed(double v, unsigned point_pos, bool saturate)
     }
 
     /* used for values inside the valid range */
-    return v * ((uint64_t)1 << point_pos);
+    return (int32_t)(v * ((uint64_t)1 << point_pos));
 }
 
 double fixed2float(uint32_t v, unsigned point_pos)
